core: move `TradeBidAsk` service to core
